# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopping Cart",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/api/products": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "19e9fc4bacd37eb1a3c54f0462345943",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "product_code": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.99
                                    },
                                    "unit_of_measurement": {
                                        "type": "string",
                                        "example": "500g"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to update product"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "f79470c2afbdb583882235160180a63d",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "product_code": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.99
                                    },
                                    "unit_of_measurement": {
                                        "type": "string",
                                        "example": "500g"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to create product"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "0fc020dce677361974d68193a62f9b18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete product"
                    }
                }
            }
        },
...
    }
}